Move OTel-Jaeger compatibility to otel_jaeger.go
Just moving the code around to make the future of tracing.NewOTel() easier to review.
